write: use bufio instead of errWriter
This simplifies the code and improves performance.

The downside is that we no longer know the number of bytes written.
(This entails an API change, which is the reason to do it now, while we can.)
But this is not a common need, and worst case,
callers can count themselves by wrapping their writer in a counting writer.