get rid of all module deps
sergi/go-diff is only used in one test, though having it as a module
dependency is still noteworthy. That will become less of a problem once
lazy module loading is widespread, but that's still at least one year
away.

Plus, we only used go-diff for a simple use case: to show a diff for a
failed diff test. We already print the got/want text, so a diff is
strictly speaking not necessary. Plus, if printing the entire text is
reasonable, it's not like the diff would matter in terms of printing
less text.

While at it, bump the go language version from 1.13 to 1.15, since Go
1.14.x and older are no longer maintained.

Fixes #23.